- MMU paging check doesn't need word addressing to work. - Fixed is_paging having the wrong result type and giving incorrect results.

+++ UniPCemu/headers/cpu/paging.h
@@ -3,7 +3,7 @@
 
 #include "headers/types.h"
 
-int is_paging();
+byte is_paging();
 uint_32 mappage(uint_32 address); //Maps a page to real memory when needed!
 byte CPU_Paging_checkPage(uint_32 address, byte readflags, byte CPL); //Do we have paging without error? userlevel=CPL usually.
